refactor(libpod): extract Client pool methods to bring mod.rs under the 500-line cap
The new connection-pool wiring pushed internal/libpod/client/mod.rs to
503 code lines, tripping line-limit on the perf/issue-1362 branch. The
pool-related methods (with_pool_size, pool_size, pool_cap,
pool_for_tests, DEFAULT_POOL_SIZE) and the impl Drop for Client block
move to internal/libpod/client/pool.rs where the pool already lives.

The public Client API is unchanged: every method that was on Client is
still reachable through the same path; the new file just owns the
construction/destruction side of the pool.

Also gates internal/libpod/client/pool/tests.rs with #![cfg(unix)] —
the file is a Unix-only integration test (it binds a UnixListener) and
failed to compile on Windows CI with 'unresolved import
`tokio::net::UnixListener`'. The pool's other unit tests still run on
Windows; only the socket-reuse integration test is Unix-only.
